alkdjalkd:

+++ install.sh
@@ -6,9 +6,14 @@
 
 # Ensure that the XDG env variables are set
 mkdir -p $HOME/.config $HOME/.cache $HOME/.local/share
-echo "export XDG_CONFIG_HOME=\$HOME/.config" | sudo tee -a /etc/zsh/zshenv > /dev/null
-echo "export XDG_CACHE_HOME=\$HOME/.cache" | sudo tee -a /etc/zsh/zshenv > /dev/null
-echo "export XDG_DATA_HOME=\$HOME/.local/share" | sudo tee -a /etc/zsh/zshenv > /dev/null
+xdg_vars=("export XDG_CONFIG_HOME=\$HOME/.config" 
+          "export XDG_CACHE_HOME=\$HOME/.cache" 
+          "export XDG_DATA_HOME=\$HOME/.localshare")
+for var in $xdg_vars; do
+    if grep -q $var /etc/zsh/zshenv; then
+        echo $var | sudo tee -a /etc/zsh/zshenv > /dev/null
+    fi
+done
 
 
 
@@ -22,7 +27,10 @@ echo "export XDG_DATA_HOME=\$HOME/.local/share" | sudo tee -a /etc/zsh/zshenv > 
 sudo pacman -S --noconfirm --needed zsh
 
 # Set ZSH dot directory
-echo "export ZDOTDIR=\$HOME/.config/zsh" | sudo tee -a /etc/zsh/zshenv > /dev/null
+$zdotdir="export ZDOTDIR=\$HOME/.config/zsh"
+if grep -q $zdotdir; then
+    echo $zdotdir | sudo tee -a /etc/zsh/zshenv > /dev/null
+fi
 
 # Ensure history file is made
 mkdir -p $HOME/.local/share/zsh
@@ -31,13 +39,17 @@ touch $HOME/.local/share/zsh/history
 
 # Install oh-my-zsh
 export ZSH=$HOME/.config/oh-my-zsh
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-$ZSH/custom}/plugins/zsh-syntax-highlighting
-git clone git://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-$ZSH/custom}/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-history-substring-search  ${ZSH_CUSTOM:-$ZSH/custom}/plugins/zsh-history-substring-search
+if [ ! -d $ZSH ]; then
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-$ZSH/custom}/plugins/zsh-syntax-highlighting
+    git clone git://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-$ZSH/custom}/plugins/zsh-autosuggestions
+    git clone https://github.com/zsh-users/zsh-history-substring-search  ${ZSH_CUSTOM:-$ZSH/custom}/plugins/zsh-history-substring-search
+fi
 
 # Link ZSH files
-ln -sf `pwd`/zsh $HOME/.config/zsh
+if [ ! -e `pwd`/zsh ]; then
+    ln -sf `pwd`/zsh $HOME/.config/zsh
+fi
 
 # Add zsh theme
 cp $HOME/.config/zsh/clean-newline.zsh-theme $HOME/.config/oh-my-zsh/themes/clean-newline.zsh-theme
@@ -57,8 +69,12 @@ sudo pacman -S --noconfirm --needed gvim xclip nodejs npm ruby python python2 st
 mkdir -p $HOME/.local/share/vim/{undo,swap,backup}
 
 # Install vim-plug
-curl -fLo ~/.local/share/vim/site/autoload/plug.vim --create-dirs \
-    https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+if [ ! -f ~/.local/share/vim/site/autoload/plug.vim ]; then
+    curl -fLo ~/.local/share/vim/site/autoload/plug.vim --create-dirs \
+        https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+fi
 
 # Link Vim files
-ln -sf `pwd`/vim $HOME/.config/vim
+if [ ! -e `pwd`/vim ]; then
+    ln -sf `pwd`/vim $HOME/.config/vim
+fi
